JBIDE-25855 - move publish controller to API
